Rename the name of the utility "backupview" -> "clickhouse_backupview".
